+++ src/core/navigation.ts
@@ -1,3 +1,0 @@
-export * from "./navigation/IToggleView";
-export * from "./navigation/nav.service";
-export * from "./navigation/directives";