+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Ignite UI for Angular Change Log
 
 All notable changes for each version of this project will be documented in this file.
+## 6.1.0
+- **Breaking changes**:
+    - Removed submodule imports. All imports are now resolved from the top level `igniteui-angular` package.
+
+## 6.0.1
+- **Breaking changes**:
+    - Removed submodule imports. All imports are now resolved from the top level `igniteui-angular` package.
+
 ## 6.0.0
 - Theming - You can now use css variables to style the component instances you include in your project.
 - Added `onDoubleClick` output to `igxGrid` to emit the double clicked cell.